mini/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-allauth==0.53.1
django-chartit==0.2.9
django-import-export==3.1.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
numpy==1.24.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
pandas==2.0.0
plotly==5.14.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tenacity==8.2.2
tzdata==2023.3
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
